// types.ts
export interface FormData {
  nickname: string;
  discord: string;
  age: string;
  about: string;
  timeOnProject: string;
  hoursDaily: string;
  activeTime: string; // Time of day they are active
  previousModExp: string;
  expectations: string;
  duties: string;
  // Quiz fields
  teamLimit: string;           
  betterPvpAllowed: string;    
  multiAccountAllowed: string; 
  recordCheckAllowed: string;  
  deanonPunishment: string;    
  weaknessPunishment: string;  
  insultModPunishment: string; 
  mentionAllowedProjects: string; 
}